+++ reader/src/ASTR/ASTR.vue
@@ -6,6 +6,7 @@ import content from './content.vue';
 import topbtn from './topbtn.vue';
 import menupage from './menupage.vue';
 import func from './func';
+import $ from 'jquery';
 
 
 export default {
@@ -14,8 +15,12 @@ export default {
             storyData: {},
             storyFile: func.storyFile,
             lang: func.l,
-            server: func.server
+            server: func.server,
+            latestUpdate: 0
         }
+    },
+    created(){
+        $.getJSON("https://raw.githubusercontent.com/050644zf/ArknightsStoryJson/main/log.json").done(s=>{console.log(s);this.latestUpdate=s['latestCommitTime']})
     },
     components:{
         Header: header,
@@ -26,6 +31,12 @@ export default {
         Menupage: menupage
     },
     methods:{
+        dateFormatter(t){
+            var d = new Date(t*1000);
+            var s = '';
+            s = d.getUTCFullYear() +'-'+(d.getUTCMonth()+1)+'-'+d.getUTCDate()+' '+d.getUTCHours()+':'+d.getUTCMinutes()+':'+d.getUTCSeconds();
+            return s;
+        }
     }
 }
 </script>
@@ -38,8 +49,21 @@ export default {
 <Menupage v-if="!storyFile"></Menupage>
 <Content v-if="storyFile"></Content>
 <Topbtn></Topbtn>
+<div class="info st">
+    <hr/>
+    Last Update (UTC): {{dateFormatter(latestUpdate)}}<br/>
+    Auto Updater scheduled at 01:30, 08:30, 11:30 UTC everyday. <br/><br/>
+    Current Status: <br/>
+    <img src="https://app.travis-ci.com/050644zf/ArknightsStoryTextReader.svg?branch=master">
+    <img src="https://github.com/050644zf/ArknightsStoryTextReader/actions/workflows/ASTRAutoUpdater.yml/badge.svg">
+</div>
 </template>
 
 <style>
-
+.info{
+    margin: 20px;
+    margin-top: 50px;
+    text-align: right;
+    color: rgba(255,255,255,0.5);
+}
 </style>