+++ reader/src/ASTR/menupage/storymenu.vue
@@ -1,7 +1,7 @@
 <template>
     <div :class="{story:true, storyFocused:focused}" :id="story.storyId" @click="loadStory(lang, story.storyTxt)">
-        <div class="storycode">{{story.storyCode}}  {{story.avgTag}} </div>
-        <div class="storyname">{{story.storyName}}</div>
+        <div class="storycode xst">{{story.storyCode}}  {{story.avgTag}} </div>
+        <div class="storyname nt">{{story.storyName}}</div>
     </div>
 </template>